+++ com.unity.ml-agents/Runtime/EnvironmentParameters.cs
@@ -25,7 +25,7 @@ namespace Unity.MLAgents
         internal EnvironmentParameters()
         {
             m_Channel = new EnvironmentParametersChannel();
-            SideChannelsManager.RegisterSideChannel(m_Channel);
+            SideChannelManager.RegisterSideChannel(m_Channel);
         }
 
         /// <summary>
@@ -64,7 +64,7 @@ namespace Unity.MLAgents
 
         internal void Dispose()
         {
-            SideChannelsManager.UnregisterSideChannel(m_Channel);
+            SideChannelManager.UnregisterSideChannel(m_Channel);
         }
     }
 }