+++ com.unity.ml-agents/Runtime/Sensors/RayPerceptionSensor.cs
@@ -469,9 +469,24 @@ namespace Unity.MLAgents.Sensors
             if (castHit)
             {
                 // Find the index of the tag of the object that was hit.
-                for (var i = 0; i < input.DetectableTags.Count; i++)
+                var numTags = input.DetectableTags?.Count ?? 0;
+                for (var i = 0; i < numTags; i++)
                 {
-                    if (hitObject.CompareTag(input.DetectableTags[i]))
+                    var tagsEqual = false;
+                    try
+                    {
+                        var tag = input.DetectableTags[i];
+                        if (!string.IsNullOrEmpty(tag))
+                        {
+                            tagsEqual = hitObject.CompareTag(tag);
+                        }
+                    }
+                    catch (UnityException e)
+                    {
+                        // If the tag is null, empty, or not a valid tag, just ignore it.
+                    }
+
+                    if (tagsEqual)
                     {
                         rayOutput.HitTaggedObject = true;
                         rayOutput.HitTagIndex = i;