+++ com.unity.ml-agents/Runtime/StatsRecorder.cs
@@ -42,7 +42,7 @@ namespace Unity.MLAgents
         internal StatsRecorder()
         {
             m_Channel = new StatsSideChannel();
-            SideChannelsManager.RegisterSideChannel(m_Channel);
+            SideChannelManager.RegisterSideChannel(m_Channel);
         }
 
         /// <summary>
@@ -65,7 +65,7 @@ namespace Unity.MLAgents
 
         internal void Dispose()
         {
-            SideChannelsManager.UnregisterSideChannel(m_Channel);
+            SideChannelManager.UnregisterSideChannel(m_Channel);
         }
     }
 }