+++ com.unity.ml-agents/CHANGELOG.md
@@ -28,10 +28,10 @@ and this project adheres to
 - Fixed an issue where RayPerceptionSensor would raise an exception when the
 list of tags was empty, or a tag in the list was invalid (unknown, null, or
 empty string). (#4155)
-- Fixed an error when setting `initialize_from` in the trainer confiiguration YAML to
-`null`. (#4175)
 
 #### ml-agents / ml-agents-envs / gym-unity (Python)
+- Fixed an error when setting `initialize_from` in the trainer confiiguration YAML to
+`null`. (#4175)
 
 ## [1.1.0-preview] - 2020-06-10
 ### Major Changes